Remove unneeded kernels (as they could be constructed by other basic ops)
