+++ tensorflow_io/core/BUILD
@@ -131,13 +131,10 @@ cc_library(
     name = "image_ops",
     srcs = [
         "kernels/font_opensans_regular.h",
-        "kernels/gif_dataset_ops.cc",
         "kernels/image_bmp_kernels.cc",
         "kernels/image_kernels.cc",
         "kernels/image_tiff_kernels.cc",
-        "kernels/tiff_dataset_ops.cc",
-        "kernels/webp_dataset_ops.cc",
-        "ops/dataset_ops.cc",
+        "kernels/image_webp_kernels.cc",
         "ops/image_ops.cc",
     ],
     copts = tf_io_copts(),
